requirements.txt
pydicom~=2.1.2
scipy>1.5.2
streamlit~=0.75.0
numpy
matplotlib~=3.3.2
Pillow~=8.1.0
opencv-python-headless

opencv-python~=4.5.1.48
open3d~=0.12.0